+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 1.15.0
 
-* Add support for Java analysis
+* Add support for Java analysis (requires [Java](https://marketplace.visualstudio.com/items?itemName=redhat.java) extension 0.56.0+)
 
 ## 1.14.0